build: code analyzer, gradle update, build file runners
- code analyzer pass to fix tons of little issues
- update default gradle version
- java language level set to 11
- gaiasky build file runners to inherit from base runner

+++ core/src/gaiasky/interfce/components/CameraComponent.java
@@ -140,7 +140,7 @@ public class CameraComponent extends GuiComponent implements IObserver {
             return false;
         });
 
-        fov = new OwnLabel((int) GlobalConf.scene.CAMERA_FOV + "°", skin, "default");
+        fov = new OwnLabel(GlobalConf.scene.CAMERA_FOV + "°", skin, "default");
 
         /** CAMERA SPEED LIMIT **/
         String[] speedLimits = new String[19];
